feat(random): add ErrNegativeLength, fix UUIDv7 32-bit overflow, and correct identifier docs
- Add ErrNegativeLength and return it from RandomBytes and RandString when n is negative, instead of panicking in make(); n == 0 still returns an empty result.
- Fix the UUIDv7 sub-millisecond precision calculation, which overflowed a 32-bit int (the intermediate product reaches 999999*4294 = 4293995706); the multiplication is now done in int64, so the package builds, vets, and tests cleanly under GOARCH=386.
- Correct the TUID64 and TUID128 type documentation to state they hold time plus random data rather than just a random value.
- Document UID64's collision ceiling (only 32 random bits per second, so birthday collisions become likely near 2^16 identifiers/second) and its decade-boundary time-ordering discontinuity.
- Simplify the UID64 return expression by dropping the redundant uint64 conversion and its //nolint:unconvert directive.
- Fix comment typos ("then" -> "the", "byiding").
- Add tests covering the negative-length paths.

+++ pkg/random/uuid.go
@@ -51,8 +51,10 @@ func (r *Rnd) UUIDv7() UUID {
 	// approximate map remaining nanoseconds in the range [0, 4095]
 	// Note:
 	// 4294 = (4095 * 2^20) / 1,000,000
-	// shifting right by 20 bits is roughly equivalent to dividing byiding by 1,048,576.
-	ns := ((now.Nanosecond() % 1e6) * 4294) >> 20
+	// shifting right by 20 bits is roughly equivalent to dividing by 1,048,576.
+	// The multiplication is done in int64 because the intermediate product can
+	// reach 999999*4294 = 4293995706, which overflows a 32-bit int.
+	ns := int((int64(now.Nanosecond()%1e6) * 4294) >> 20)
 
 	// ver:
 	// 4-bit version field set to 0b0111 (7)
